+++ data/doc.go
@@ -1,0 +1,2 @@
+// Package data provides implementations for data structures, like graphs.
+package data
