+++ data/graph_test.go
@@ -1,4 +1,4 @@
-package graph
+package data
 
 import (
 	"testing"
