--- graph/doc.go
@@ -1,2 +1,0 @@
-// Package graph provides implementations for both directed and undirected graphs, using adjacency lists.
-package graph
